+++ package.json
@@ -101,7 +101,6 @@
     "@toruslabs/customauth-react-native-sdk": "^1.0.1",
     "@toruslabs/torus-direct-web-sdk": "5.0.2",
     "@tradle/react-native-http": "^2.0.1",
-    "FormData": "^0.10.1",
     "abi-decoder": "^1.2.0",
     "abortcontroller-polyfill": "^1.7.3",
     "amplitude-js": "^8.8.0",
@@ -120,8 +119,10 @@
     "dotenv-override": "^5.0.1",
     "es6-promisify": "^6.1.0",
     "eth-multicall": "^1.4.0",
+    "ethereumjs-tx": "^2.1.2",
     "eventemitter3": "^4.0.7",
     "events": "^1.1.1",
+    "FormData": "^0.10.1",
     "hdkey": "^1.1.2",
     "hex-to-rgba": "^2.0.1",
     "https-browserify": "0.0.1",
@@ -240,6 +241,7 @@
     "vm-browserify": "^1.1.2",
     "web3": "^1.5.1",
     "web3-core": "^1.5.2",
+    "web3-provider-engine": "^16.0.4",
     "web3-utils": "^1.5.2",
     "webpack-bundle-analyzer": "^4.5.0",
     "websocket": "^1.0.31"
